+++ extra_foam/gui/windows/correlation_w.py
@@ -168,7 +168,7 @@ class CorrelationWindow(_AbstractPlotWindow):
         self.updateMetaData()
 
         self.resize(self._TOTAL_W, self._TOTAL_H)
-        self.setMinimumSize(0.6*self._TOTAL_W, 0.6*self._TOTAL_H)
+        self.setMinimumSize(int(0.6 * self._TOTAL_W), int(0.6 * self._TOTAL_H))
 
         self.update()
